src: modify int to uint16_t fori `HostPort`

+++ src/inspector_js_api.cc
@@ -286,7 +286,7 @@ void Open(const FunctionCallbackInfo<Value>& args) {
     uint32_t port = args[0].As<Uint32>()->Value();
     CHECK_LE(port, std::numeric_limits<uint16_t>::max());
     ExclusiveAccess<HostPort>::Scoped host_port(agent->host_port());
-    host_port->set_port(static_cast<int>(port));
+    host_port->set_port(static_cast<uint16_t>(port));
   }
 
   if (args.Length() > 1 && args[1]->IsString()) {
